Removes hidden pages from route list
