+++ src/router/index.js
@@ -1,27 +1,6 @@
 // import { h, resolveComponent } from 'vue'
 import { createRouter, createWebHashHistory } from 'vue-router'
-
-import DefaultLayout from '@/layouts/DefaultLayout'
-
-const routes = [
-  {
-    path: '/',
-    name: 'Home',
-    component: DefaultLayout,
-    redirect: '/dashboard',
-    children: [
-      {
-        path: '/dashboard',
-        name: 'Dashboard',
-        // route level code-splitting
-        // this generates a separate chunk (about.[hash].js) for this route
-        // which is lazy-loaded when the route is visited.
-        component: () =>
-          import(/* webpackChunkName: "dashboard" */ '@/views/Dashboard.vue'),
-      },
-    ],
-  },
-]
+import routes from './routes'
 
 const router = createRouter({
   history: createWebHashHistory(process.env.BASE_URL),
